pyproject.toml
[project]
name = "ai-knowledgebased-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "chromadb>=1.5.8",
    "fastapi[standard]>=0.136.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.14",
    "langchain-text-splitters>=1.1.2",
    "openai>=2.32.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.13.2",
    "pyjwt>=2.12.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "tiktoken>=0.12.0",
    "unstructured[all-docs]>=0.22.22",
]

[tool.fastapi]
dev = "app/main.py"
